+++ resources/lang/en.json
@@ -619,6 +619,8 @@
     "name_cull_desc": "Hide names smaller than this size",
     "name_cull_label": "Minimum name size",
     "name_scale_label": "Name Scale",
+    "naval_hover_highlight_desc": "Highlight the owner of naval units when hovering over them.",
+    "naval_hover_highlight_label": "Naval Hover Highlight",
     "nuke_color_desc": "Color of the fallout tint left on territory after a nuke.",
     "nuke_color_label": "Nuke fallout color",
     "ocean_color_desc": "Base color of ocean.",
@@ -1550,8 +1552,6 @@
     "move_right_desc": "Move the camera to the right",
     "move_up": "Move Camera Up",
     "move_up_desc": "Move the camera upward",
-    "naval_hover_highlight_desc": "Highlight the owner of naval units when hovering over them.",
-    "naval_hover_highlight_label": "Naval Hover Highlight",
     "off": "Off",
     "on": "On",
     "pause_game": "Pause",
