+++ src/client/render/gl/render-settings.json
@@ -72,6 +72,7 @@
     "lightRadiusMultiplier": 1
   },
   "mapOverlay": {
+    "navalHighlight": false,
     "trailAlpha": 0.588,
     "defenseCheckerDarken": 0.7,
     "territoryDefenseDarken": 0.85,
