+++ src/core/game/UserSettings.ts
@@ -175,10 +175,6 @@ export class UserSettings {
     return this.getBool("settings.goToPlayer", true);
   }
 
-  navalHoverHighlight() {
-    return this.getBool("settings.navalHoverHighlight", false);
-  }
-
   attackingTroopsOverlay() {
     return this.getBool("settings.attackingTroopsOverlay", true);
   }
@@ -246,10 +242,6 @@ export class UserSettings {
 
   toggleGoToPlayer() {
     this.setBool("settings.goToPlayer", !this.goToPlayer());
-  }
-
-  toggleNavalHoverHiglight() {
-    this.setBool("settings.navalHoverHighlight", !this.navalHoverHighlight());
   }
 
   // For development only. Used for testing patterns, set in the console manually.
